Remove binary files from the repo
Binary files are not supposed to be distributed with the source code. It
is just needless clutter that does not belong in a git repository.

Please stop committing binary files to the repository. The files that I
deleted literally take up more than 15x the space that every other file
combined uses. Every time someone clones the repository, they have to
download it. Every time one of these files is modified, its size is
effectively doubled in the repo because git keeps track of the history,
but it's not designed for binary files so it creates a new copy of it.

Do not commit binary files to git repositories.